Add new modeule for rate aggregation
